% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadf.R
\name{rtadf}
\alias{rtadf}
\title{rtadf}
\usage{
rtadf(y, r0, test = c("adf", "sadf", "gsadf"), type = c("none", "drift",
  "trend"), lags = 1, selectlags = c("Fixed", "AIC", "BIC"))
}
\arguments{
\item{y}{Vector to be tested for an explosive root.}

\item{r0}{Minimal Window size.}

\item{test}{Test type, either "adf", "sadf" of "gsadf".}

\item{type}{Deterministic terms, either, either "none", "drift" or "trend".}

\item{lags}{Number of lags for y to be included.}

\item{selectlags}{Lag selection can be achieved according to the Akaike "AIC"
or the Bayes "BIC" information criteria. The maximum number of lags
considered is set by lags. The default is to use a "fixed" lag length set
by lags.}
}
\value{
List with test statistic, date-stamping sequence
}
\description{
Calculates recursive and non-recursive right-tailed unit root tests statistics.
}
\examples{
y  <- rnorm(100)
r0 <- round(t*(0.01+1.8/sqrt(t))) #minimal window size
rtadf(y, r0, test ="sadf")
}
