% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadfCval.R
\name{rtadfCval}
\alias{rtadfCval}
\title{Returns critical values based on a response surface function approximation}
\usage{
rtadfCval(t, pval, testType)
}
\arguments{
\item{t}{Number of observations (i.e., length of the sample)}

\item{pval}{Significance level (in percent)}

\item{testType}{Test type, either "adf", "sadf" of "gsadf".}
}
\value{
Numeric, critical value at the user-specified significance level.
}
\description{
\code{rtadfCval} approximate critical values for the SADF test (Phillips,Wu
and Yu, 2011) using the MacKinnon (1996) response surface function approach
(Caspi, 2018).
}
\examples{
cv <- rtadfCval(t = 100,  pval = 0.95, testType = "sadf")
}
\references{
Caspi, I. (2018) Empirical Distribution Functions for
  Right-Tailed Unit Root tests for Exuberance. Unpublished mimeo.

  MacKinnon, J. G. (1996). Numerical distribution functions for unit root and
  cointegration tests. \emph{Journal of Applied Econometrics}, 11(6):601–618.

  Phillips, P. C. B., Wu, Y., & Yu, J. (2011). Explosive Behavior in the
  1990s Nasdaq: When Did Exuberance Escalate Asset Values?,
  \emph{International Economic Review}, 201(1), 201--226.

  Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for multiple bubbles:
  Historical episodes of exuberance and collapse in the S&P 500.
  \emph{International Economic Review}, 56(4), 1034--1078.
}
