% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadfPval.R
\name{rtadfPval}
\alias{rtadfPval}
\title{Returns p-values based on a response surface function approximation}
\usage{
rtadfPval(t, testStat, testType)
}
\arguments{
\item{t}{Number of observations (i.e., length of the sample)}

\item{testStat}{Test statistic}

\item{testType}{Test type, either "adf", "sadf" of "gsadf".}
}
\value{
Numeric, p-value that corresponds to a user-specified test statistic.
}
\description{
\code{rtadfPval} approximate p-values for the SADF test (Phillips,Wu and Yu,
2011) using the MacKinnon (1996) response surface function approach (Caspi,
2018).
}
\examples{
pval <- rtadfPval(t = 100,  testStat = 1.5, testType = "sadf")
}
\references{
Caspi, I. (2018) Empirical Distribution Functions for
  Right-Tailed Unit Root tests for Exuberance. Unpublished mimeo.

  MacKinnon, J. G. (1996). Numerical distribution functions for unit root and
  cointegration tests. \emph{Journal of Applied Econometrics}, 11(6):601–618.

  Phillips, P. C. B., Wu, Y., & Yu, J. (2011). Explosive Behavior in the
  1990s Nasdaq: When Did Exuberance Escalate Asset Values?,
  \emph{International Economic Review}, 201(1), 201--226.

  Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for multiple bubbles:
  Historical episodes of exuberance and collapse in the S&P 500.
  \emph{International Economic Review}, 56(4), 1034--1078.
}
