% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadfSim.R
\name{rtadfSim}
\alias{rtadfSim}
\title{Simulate critical values and date-stamping threshold sequence}
\usage{
rtadfSim(t, r0, nrep = 1000, test = c("adf", "sadf", "gsadf"))
}
\arguments{
\item{t}{Number of observations (i.e., length of the sample)}

\item{r0}{Minimal window size}

\item{nrep}{Number of replications}

\item{test}{Test type, either "adf", "sadf" of "gsadf".}
}
\value{
List with 90, 95, 99 precent critical values and a threshold series
  for the date-stamping procedure
}
\description{
Simulates critical values and date-stamping sequences for the right-tailed
ADF, SADF and GSADF tests.
}
\examples{
cv <- rtadfSim(t = 100, r0 = 10, nrep = 1000, test = "sadf")
}
