% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadfSimPar.R
\name{rtadfSimPar}
\alias{rtadfSimPar}
\title{Simulate critical values and date-stamping threshold sequence in parallel}
\usage{
rtadfSimPar(t, nrep, testType)
}
\arguments{
\item{t}{Number of observations (i.e., length of the sample)}

\item{nrep}{Number of replications}

\item{testType}{Test type, either "adf", "sadf" of "gsadf".}

\item{r0}{Minimal window size}
}
\value{
List with 90, 95, 99 precent critical values and a threshold series
  for the date-stamping procedure
}
\description{
Simulates in parallel critical values and date-stamping sequences for the
right-tailed ADF, SADF and GSADF tests.
}
\examples{
cv <- cvSimPar(100, 10, 1000, testType = "sadf")
}
