% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp}
\alias{snp}
\title{S&P 500 price to dividend ratio index.}
\format{A data frame with 1680 monthly observations (January 1871 - December
  2010)}
\usage{
snp
}
\description{
A dataset containing an index of the S&P 500 price to dividend ratio
}
\keyword{datasets}
