% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadfSimPar.R
\name{rtadfSimPar}
\alias{rtadfSimPar}
\title{Simulate critical values and date-stamping threshold sequence in parallel}
\usage{
rtadfSimPar(t, r0, nrep = 1000, test = c("adf", "sadf", "gsadf"))
}
\arguments{
\item{t}{Number of observations (i.e., length of the sample)}

\item{r0}{Minimal window size}

\item{nrep}{Number of replications}

\item{test}{Test type, either "adf", "sadf" of "gsadf".}
}
\value{
List with 90, 95, 99 precent critical values and a threshold series
  for the date-stamping procedure
}
\description{
Simulates in parallel critical values and date-stamping sequences for the
right-tailed ADF, SADF and GSADF tests.
}
\examples{
cv <- rtadfSimPar(t = 100, r0 = 10, nrep = 100, test = "sadf")
}
\references{
Phillips, P. C. B., Wu, Y., & Yu, J. (2011). Explosive Behavior
in the 1990s Nasdaq: When Did Exuberance Escalate Asset Values?,
\emph{International Economic Review}, 201(1), 201--226.

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for multiple bubbles:
Historical episodes of exuberance and collapse in the S&P 500.
\emph{International Economic Review}, 56(4), 1034--1078.
}
